add additional BaseFedJob layer
