+++ nvflare/app_opt/pt/job_config/fed_avg.py
@@ -16,10 +16,10 @@ from typing import List, Optional
 import torch.nn as nn
 
 from nvflare.app_common.workflows.fedavg import FedAvg
-from nvflare.app_opt.pt.job_config.base_fed_job import BaseFedJob
+from nvflare.app_opt.pt.job_config.base_fed_job import PTBaseFedJob
 
 
-class FedAvgJob(BaseFedJob):
+class FedAvgJob(PTBaseFedJob):
     def __init__(
         self,
         initial_model: nn.Module,
