+++ nvflare/app_opt/tf/job_config/fed_avg.py
@@ -16,10 +16,10 @@ from typing import List, Optional
 import tensorflow as tf
 
 from nvflare.app_common.workflows.fedavg import FedAvg
-from nvflare.app_opt.tf.job_config.base_fed_job import BaseFedJob
+from nvflare.app_opt.tf.job_config.base_fed_job import TFBaseFedJob
 
 
-class FedAvgJob(BaseFedJob):
+class FedAvgJob(TFBaseFedJob):
     def __init__(
         self,
         initial_model: tf.keras.Model,
